// src/app/api/model/userEdit.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RolDB } from './rolDB';


export interface UserEdit { 
    id?: number;
    roles?: Array<RolDB>;
    name?: string;
    email?: string;
    enabled?: boolean;
}